pyproject.toml
[project]
name = "beanqueue"
version = "1.1.9"
description = "BeanQueue or BQ for short, PostgreSQL SKIP LOCK and SQLAlchemy based worker queue library"
authors = [{ name = "Fang-Pen Lin", email = "fangpen@launchplatform.com" }]
requires-python = "~=3.11"
readme = "README.md"
license = "MIT"
dependencies = [
    "sqlalchemy>=2.0.30,<3",
    "venusian>=3.1.0,<4",
    "click>=8.1.7,<9",
    "pydantic-settings>=2.2.1,<3",
    "blinker>=1.8.2,<2",
    "rich>=13.7.1,<14",
]

[project.scripts]
bq = "bq.cmds.main:cli"

[dependency-groups]
dev = [
    "psycopg2-binary>=2.9.9,<3",
    "pytest-factoryboy>=2.7.0,<3",
]

[tool.hatch.build.targets.sdist]
include = ["bq"]

[tool.hatch.build.targets.wheel]
include = ["bq"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
